+++ Web/GlobalExceptionHandler.cs
@@ -3,14 +3,18 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Web;
 
-#pragma warning disable CA181, CA1515// Avoid uninstantiated internal classes
 public sealed class GlobalExceptionHandler(ILogger<GlobalExceptionHandler> logger) : IExceptionHandler
-#pragma warning restore CA1812, CA1515 // Avoid uninstantiated internal classes
 {
     public async ValueTask<bool> TryHandleAsync(HttpContext httpContext, Exception exception, CancellationToken cancellationToken)
     {
         if (exception is null || httpContext is null)
         {
+            return false;
+        }
+
+        if (httpContext.Response.HasStarted)
+        {
+            logger.LogWarning("Response already started; skipping exception handling.");
             return false;
         }
 
@@ -23,7 +27,7 @@ public sealed class GlobalExceptionHandler(ILogger<GlobalExceptionHandler> logge
         };
 
         httpContext.Response.StatusCode = problemDetails.Status.Value;
-
+        httpContext.Response.ContentType = "application/problem+json";
         await httpContext.Response.WriteAsJsonAsync(problemDetails, cancellationToken);
 
         return true;
