+++ Web/StartupInfo.cs
@@ -10,7 +10,7 @@ internal static class StartupInfo
 {
     private const double Mebi = 1024 * 1024;
     private const double Gibi = Mebi * 1024;
-    private const string Mcm_Prodcut = "MCM";
+    private const string Mcm_Product = "MCM";
 
     internal static void Print()
     {
@@ -41,7 +41,7 @@ internal static class StartupInfo
         foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies())
         {
             var productAttribute = assembly.GetCustomAttribute<AssemblyProductAttribute>();
-            if (productAttribute != null && productAttribute.Product == Mcm_Prodcut)
+            if (productAttribute != null && productAttribute.Product == Mcm_Product)
             {
                 Log.Information("Assembly Versions: {@AssemblyName} - {@AssemblyVersion}", assembly.GetName().Name, assembly.GetName().Version!.ToString());
             }
