+++ Web/Validators/LibraryValidator.cs
@@ -6,13 +6,24 @@ public class LibraryValidator : AbstractValidator<LibraryUiDto>
     public LibraryValidator()
     {
         const int maxNameLength = 100;
-        RuleFor(x => x.Name).NotEmpty().WithMessage("Name is required.").MaximumLength(maxNameLength).WithMessage("Name must not exceed 100 characters.");
-        RuleFor(x => x.RelativePath).NotEmpty().WithMessage("RelativePath is required.").MaximumLength(maxNameLength).WithMessage("RelativePath must not exceed 200 characters.");
+        RuleFor(x => x.Name).NotEmpty().WithMessage("Name is required.").MaximumLength(maxNameLength).WithMessage($"Name must not exceed {maxNameLength} characters.");
+        RuleFor(x => x.RelativePath).NotEmpty().WithMessage("RelativePath is required.").MaximumLength(maxNameLength).WithMessage($"RelativePath must not exceed {maxNameLength} characters.");
     }
 
     public Func<object, string, Task<IEnumerable<string>>> ValidateValue => async (model, propertyName) =>
     {
-        var result = await ValidateAsync(ValidationContext<LibraryUiDto>.CreateWithOptions((LibraryUiDto)model, x => x.IncludeProperties(propertyName)));
+        if (model is not LibraryUiDto dto)
+        {
+            return new[] { "Invalid model for validation." };
+        }
+
+        if (string.IsNullOrWhiteSpace(propertyName))
+        {
+            return new[] { "Property name is required." };
+        }
+
+        var ctx = ValidationContext<LibraryUiDto>.CreateWithOptions(dto, x => x.IncludeProperties(propertyName));
+        var result = await ValidateAsync(ctx);
         if (result.IsValid)
         {
             return Array.Empty<string>();
